satellite/metabase: replace ListObjects implementation
ListObjects implementation turns out to be slow with cockroach so for
now we decided to replace it with well known objects iterator and just
filter latest objects. We are also changing ListObjects to not support
listing pending objects and all versions.

We keep old implementation as the plan is to improve it but we need time
to do that.

Change-Id: Icd6e629c9ab8519744843d91a985404d92dd1594