storagenode/{pieces,blobstore}: save-state-resume feature for GC filewalker
The bloomfilter is now stored on the disk from this change
https://review.dev.storj.io/c/storj/storj/+/12651, so we don't lose
the bloomfilter on restart. But that means that when the bloomfilter
on the disk is loaded and added to the retain queue, it will start
scanning the directories all over again. This will continuously be the
behavior till the node is able to complete a full scan which might never
happen on a large node.

In this patch, we make sure the node can resume from where it left off.
We read the two-letter directory names (or prefixes) from the satellite
blobs folder and sort them alphabetically before scanning.
During the scan, the last scanned prefix is stored in the db.

Updates https://github.com/storj/storj/issues/6708

Change-Id: Icb32cdc7dd49ef8ce44f6d771e4e33045078ed55