satellite/console: Add Cache-Control headers
Set headers on the UI and API to not store cache for requests.
Static content can still be cached by the browser. Caching for one hour
also remains on a couple usage endpoints that require authentication.
By default, Cache-Control=no-store is now set on all requests. It must
be overridden in cases where caching is desireable.

https://github.com/storj/storj-private/issues/294

Change-Id: Icf2d466b1bce49f8917126028816844e43802faf