storagenode/pieces: trash pieces immediately in GC Filewalker
Previously, GCfilewalker only finds the trash pieces and
returns them for the retain service to trash them. This
behavious has cons: if the GC filewalker fails or the node
is stopped before the filewalker completes, no piece will
be trashed and the found trash pieces are lost.
This is a big problem for larger nodes that might not
complete GC filewalker earlier before it is auto-updated.

In the patch, the filewalker trashes the piece immediately
when found. For the lazyfilewalker, when a trash piece is
found, the pieceID is piped to the main process so that it
will be trashed. Note that we do not let the subprocess to
trash it because we want to be able to update the space usage
cache. Once we are done with the lazyfilwaker optimization,
we might come back to this and let the subprocess remove
the trash piece by itself.

Updates https://github.com/storj/storj/issues/6803

Change-Id: I40b52e3e940a398ecc46d0d1c448f098173f5f99