web/satellite: added trial expiration banner and dialog
Added dialog which pops up on login if user's trial is expired or close to be expired.
Added banner to project dashboard which is shown if user's trial is expired or close to be expired.
Added banner to upgrade dialog to inform user that upgrading ends free trial immediately.

TODO: restrict some actions if user is expired.

Issue:
https://github.com/storj/storj-private/issues/604

Change-Id: I3504319fdbf43456884ecd18e130a9957f91f315