web/satellite: Fix project limit UI bug
On the "all projects" dashboard, projects that a user was invited to,
but did not own, were included in the "total projects" count when
determining whether to allow the user to create a project, or prompt
them to upgrade or request a limit increase.

This change fixes the problem by only counting a user's owned projects
when deciding whether to allow the user to make a "create project"
request.

https://github.com/storj/storj/issues/6864

Change-Id: I6c86f659d400079e0bafdfa49c4a7f052970b04f