satellite/analytics: add event Arrived From Source
Add new analytics event, Arrived From Source. This event is triggered
during the setup phase of the frontend app if the 'source' query
parameter exists. On the backend we define a map of allowed sources
based on constants. This will allow us to identify what brought the user
to the satellite console; in this case, email links.

updates: https://github.com/storj/storj-private/issues/609

Change-Id: I620450c3db2986564d8344cd7e30be5392c9c053