satellite/metainfo: Log internal error and warning
Deleting a bucket with data without having read and list permissions
isn't allowed, but we return a successful response to avoid attackers to
figure out if a bucket exists.

We never return an error to the clients under those circumstances, but
we want to see them them in the logs.

Change-Id: I1bb0dd3fa0c3534b17d98095c002032738eea103