storagenode/pieces: save-state-resume feature for used space filewalker
This change adds a new feature to the used space filewalker that
allows it to resume from the last point it stopped.
This is useful for when the node is restarted, so it doesn't have to
start from the beginning.

To give a little overview of the process:
1. Cache service stores the current used space before the piece scan starts
   in the db. This is so that we can check the difference and know if any pieces
   were added, trashed or deleted.
2. Cache service calls SpaceUsedTotalAndBySatellite() on the store which gets
   all the available blobs namespaces, and then sequentially starts a
   filewalker/lazyfilewalker for each namespace (satellite).
3. The filewalker stores the used space per prefix and keep track of the last
   prefix after each prefix scan. And when completed without interruption,
   returns the total usage for the satellite/namespace to the store.
   The store then marks that the satellite has completed the scan and starts
   a filewalker/lazyfilewalker for the next namespace/satellite. When completed
   for all satellites, store returns the total to the cache service and then
   removes all progress data

What happens if the node shuts down during this process?
TLDR; progress isn't lost, and the filewalker starts the calculation from the last
prefix, sums the total usage for all scanned prefixes and returns it.

4. If the node restarts, we go back to (1) but cache service checks if we had already
   saved a used-space before a previous scan which was incomplete and load that and then
   we are back to (2).
5. The store checks if scans were completed for any of the namespaces and then loads
   the usage from the db and skips filewalker for them, and then starts the
   filewalker/lazyfilwalker for the namespaces with incomplete scan.
6. Filewalker checks progress for last scan prefix and resume from there.

(4),(5),(6) repeats on every restart till scan is completed for all namespaces and the
cache persists the data to the db.

This isn't perfect but the beginning of providing a better solution to used-space
calculation

Updates https://github.com/storj/storj/issues/6708

Change-Id: I730aa6cef66d3009a34fd81aa8b6461b65623812