storagenode/pieces/lazyfilewalker: fix gc-filewalker
This change fixes the GC filewalker subprocess output not being piped
to the trashHandler.

Previously, the `stdout` of the subprocess was set to an
`io.ReadWriter`.
This caused issues because the `trashHandler` (which implements
`io.ReadWriter`) wasn't actively reading from the pipe, leading
to data being buffered and not processed.

This commit changes the `stdout` type to `io.Writer`.
This clarifies the intent and ensures the subprocess can write
its output without relying on the `trashHandler` to read from
the pipe.

Updates https://github.com/storj/storj-private/issues/666

Change-Id: I6139018e0eaae5c0c7bee0dfdd5bbba69b948a99