satellite/metabase: accept spanner as a first class db implementation
Earlier we tried to use spanner as an overlay option to make the development easier. When Spanner version is not yet implemented, the Cocorkach was used.

It turned to be too complex, so we are moving back to supporting Spanner as first class database (with `spanner://`)

To make the development easier we already implemented a special flag for metabase test to execute test on Spanner, in addition to Postgres and Cocrocach.

Change-Id: I38ca1cdaf09a7c86a82c18d0e75a5937678853cf