satellite/metainfo: keep track of and use upload successes
this change introduces a mechanism for the satellite
to keep track of which nodes successfully received
pieces during segment uploads and prefer those nodes
during the selection process. it does so by
introducing a success tracker type to keep the
information and implements a pairwise power-of-2
choices style selection mechanism during segment
uploads.

the success tracker is a sync.Map of node id to an
array of counters. each counter keeps track of the
number of successes and the total number of attempts
as two uint32s packed into a single uint64 for both
efficiency and atomicity. in the common case, there
should be no additional contention to increment or
get the counter by avoiding the use of mutexes.

it is an array of counters so that we can have a
sliding window that decays the information over
time. the tracker keeps track of a generation
and bumps it periodically to cause new Increment
calls to go to the next generation (index in the
array of counters). there is no coordination
between the bumping and any existing Increment
calls, so it's possible to write to the generation
after it has been cleared, but this is not a
correctness issue: some extra information will
last slightly longer than it should.

the endpoint can be configured with a list of ids
of uplinks to allow them to have their own success
tracker (think gateways in europe may prefer nodes
around europe as compared to gateways in the us).
if the uplink is not in the list, it uses a global
success tracker.

Change-Id: I6af02aea1d09f0d03c97e8b759882a23a713391b