storagenode/orders: Open "writable unsent orders file" less
Prior to this change, every time a new order limit needed to be written
to disk, the "unsent orders" file was opened, the order was written, and
the "unsent orders" file was closed.

After this change, the "unsent orders" file is opened, and stays open
for all order limits until an order arrives that should be inserted into
a different file (e.g. an hour has elapsed). At this point, the old
"active unsent orders" file is closed and replaced with the next one.

A lifecycle item is added to the storagenode peer for the file store, so
that an open "active unsent orders" file can be closed when the
storagenode shuts down.

https: //github.com/storj/storj-private/issues/716
Change-Id: I1fb00431734932634c6b9573e530a22936cd9815