shared/dbutil: spannerutil.Int utility function
This will make it simpler to pass output locations to our Spanner
library, which apparently only likes int64 out of all Go's integer
types.

In general, it is preferable to add EncodeSpanner/DecodeSpanner meethods
to our specialized int types, but this can be used for types we don't
own or otherwise can't put methods on (in particular, int32).

Change-Id: Ibdb64e4483e1de32fa13f2f4b464887fadd6e48c