storagenode/blobstore: remove preallocate size
we weren't actually preallocating and instead
just calling ftruncate, which doesn't do anything
besides setting the file size in the inode. under
filesystems with delayed allocation, this is just
a metadata update that we eventually undo during
commit, so it's useless work that makes things
slower. instead, never call ftruncate and let
the writes automatically update the size as
usual.

Change-Id: I9da9821d77cef92d68913201e1978ed022e95225