blobstore/filestore: cache Info for a second
this shows up in speed profiles as a hot spot
on windows. in a test of 50k pieces, it took
67 seconds out of 145 seconds, leaving 78 total
seconds so we should see a reduction of 78/50000
or 0.1% of the time, or ~600ms. should be good
enough and still maintain enough freshness for
it to do a good enough job enforcing the available
space on the piece upload.

Change-Id: Iaaff8b7739c52bcb5313c283c5aa3750c9f9139f