storagenode/orders: Log close err instead of return
We know of at least one situation where attempting to close this file
has prevented opening the new unsent orders file, because the old file
was already closed.

It is still unclear to me how this state is introduced in the first
place, but this fix should work regardless. Don't return the close error
- just log it.

Change-Id: Iae8020c51d7c4b116eb851a6901f5b2a0dc70fb5