cherry-pick for v1.104.6-rc2 (#6975)
* satellite/metabase: new CommitInlineObject method

This method will help optimize uploading objects with single inline
segment. We are batching all needed requests (begin object, make
inline segment, commit object) so we can do full object upload as a
single transaction and with minimal amount of SQL queries. This change
will be later used by metainfo endpoints.

Change-Id: I1949440a9bf82cda0c3ac3830d5083d0c24c2bc6

* satellite/metainfo: commit inline object as single request

This is using new metabase CommitInlineObject method to optimize
uploading objects with single inline segment. If we detect that all
needed requests are batched together then we are calling single method
to cover begin object, make inline segment and commit object.
After change with upload we will have single transaction with minimal
number of SQL queries:
* find highest available version
   * (if not versioned) delete previous object and it's segments
* insert object
* insert segment

As next improvement we could combine inserting object and segment into
single query but I would prefer first to test them separately.

This optimization is behind feature flag.

Change-Id: Ia9994a8c0e3850ddbac150d9cb62d28d16dd86fd

* satellite/metainfo: keep track of and use upload successes

this change introduces a mechanism for the satellite
to keep track of which nodes successfully received
pieces during segment uploads and prefer those nodes
during the selection process. it does so by
introducing a success tracker type to keep the
information and implements a pairwise power-of-2
choices style selection mechanism during segment
uploads.

the success tracker is a sync.Map of node id to an
array of counters. each counter keeps track of the
number of successes and the total number of attempts
as two uint32s packed into a single uint64 for both
efficiency and atomicity. in the common case, there
should be no additional contention to increment or
get the counter by avoiding the use of mutexes.

it is an array of counters so that we can have a
sliding window that decays the information over
time. the tracker keeps track of a generation
and bumps it periodically to cause new Increment
calls to go to the next generation (index in the
array of counters). there is no coordination
between the bumping and any existing Increment
calls, so it's possible to write to the generation
after it has been cleared, but this is not a
correctness issue: some extra information will
last slightly longer than it should.

the endpoint can be configured with a list of ids
of uplinks to allow them to have their own success
tracker (think gateways in europe may prefer nodes
around europe as compared to gateways in the us).
if the uplink is not in the list, it uses a global
success tracker.

Change-Id: I6af02aea1d09f0d03c97e8b759882a23a713391b

---------

Co-authored-by: Michal Niewrzal <michal@storj.io>
Co-authored-by: Jeff Wendling <jeff@storj.io>