satellite/kms: add support for multiple kms keys
Rework kms service to support multiple keys. If we ever want to rotate
the master key, we still need to be able decrypt passphrases which were
encrypted using the previous key until they are migrated. To accomplish
this, the --kms.secret-version and --kms.secret-checksum configs are
moved into a map, --kms.key-infos, where each key is identified by an
integer. This integer key ID identifies which key was used to encrypt
the project passphrase and it is stored in the projects table. Also,
instead of mocking the entire kms.secretService, this change mocks the
secretmanager client used by the gsmService, so more of the logic in
gsmService is tested.

issue: https://github.com/storj/storj/issues/6955

Change-Id: Ic7b37c42ef68246939170d0909823a76c450416a