satellite/metabase: add new precommit benchmark and commit test
Add benchmark test to help in comparing the performance of unversioned
object precommits: precommit for an object that requires delete
(overwrite) vs. for an object that does not require delete (no
overwrite). With variations for different precommit delete modes.
Also add a new precommit mode varied test for commit object.

Change-Id: Ib6aafdc8be44fb4dc153bf4464ebf3f58148e83b