private/mud: handle case where an interface type is required
Previously, if an interface type was required by an execution with a mud
ball, the right instance would be exposed, but that instance's Run() and
Close() methods (if any) were not correctly registered. This is the
reason why builds have not been cleaning up their temporary databases.

Using t.MethodByName("Foo") on a reflect.Type object, when that type is
an interface type, returns a reflect.Method with a nil Func member,
which previously caused registerFunc[T]() to skip it.

Now, we will register a function which will find the appropriate "real"
method to call once an instance of the interface is available.

Change-Id: I5a6b6750d27d28353dc203b3aead15d2b48cc923