satellite/nodeselection: Increase likelihood to select unvetted
We want to increase the likelihood to select unvetted nodes.

Before these changes the unvetted selector was not choosing any selected
node if the fraction of the number of nodes to select resulted in less
than 1.

With these changes, if the number of nodes to select results in less
than 1, it randomly determine if one unvetted nodes is selected or not.
The random likelihood to select 1 node or none varies with the decimal
value.

This commit also refactor a test with several sub-cases inside of scopes
to sub-tests.

Change-Id: Id938b45d43e549c1ada33b319da99ca598a93dc7