satellite/metabase: add UpdateTableStats() method to DB
This supplies a new database method, UpdateTableStats(). Most
importantly, this removes some uses of UnderlyingTagSQL().

I've also taken the opportunity to fill in code so that statistics also
work on PostgreSQL. We don't use PostgreSQL much with the satellite, but
it may be used more for SPC deployments. Anyway, PostgreSQL does support
and collect table statistics, so maybe we should be able to make use of
them.

Change-Id: Iae2abfbd29be5caf5d95cd94721f4fb31115b233