satellite/payments: update invoicing logic to handle accounts marked for deletion
During the invoice generation process, check if the user is marked_for_deletion and if status_updated_at exists within the billing period of invoicing.
If so, set the upper limit on the billing period to status_updated_at, so they aren't billed for stored data since the time they opted for deletion.
Once the invoice is finalized, set final_invoice_generated to true.

Issue:
https://github.com/storj/storj/issues/6965

Change-Id: I3bc6f9bc86ec034ab6ea7b80305a02f61c55c0e4