storagenode/storagenodedb: buffer up GC Filewalker progress storage
Previously, the GC filewalker stores the progress after completing
a full scan of each prefix. This consequently could lead to the
database lock issues, if there are multiple gc filewalkers running
as reported on the forum.

This patch makes sure the db is only updated after the
filewalker completes scanning a specified number of
prefixes (currently 5; hardcoded).

https://forum.storj.io/t/error-lazyfilewalker-gc-filewalker-subprocess-failed-to-save-progress-in-the-database/26407

Change-Id: I26a0cbb778b74095f2b1c83221896d856f1052dc