web/satellite: improve exceed bandwidth limit error handling
This change improves the way the UI behaves when bandwidth limit is
exceeded;
Media previews will stop loading on this error and show a placeholder.
Text previews will now show a placeholder on this error.
Download will not be allowed if the desired file is larger than the
remaining bandwidth.

Issue: https://github.com/storj/storj/issues/7013

Change-Id: Ib5a45fdf5a599cb5c9fcbec5385c908e4251193f