satellite/metabase: update GetTableStats for postgresql
The n_ins_since_vacuum column is not present in older Postgresql
versions (such as version 12.3 which is used in one of our Jenkins
setups). Furthermore, its use was not sufficient to solve the problem,
and we ended up excusing postgresql from the relevant test instead
in commit cd7e49a3.

Change-Id: I98e32d9f88fabd499e4df395c489560860c5402d