satellite/nodeselection: support choiceofn node selection on the download side
we don't have a way to track download successes like we do with upload
successes, but the upload success tracker is maybe still kind of
useful as a proxy for understanding node loadedness. if we assume the
upload success tracker does correlate with download performance, then
this change might not be so wacky.

this change adds a new type of download selector. its job is to winnow
down the total number of nodes for a given segment down to the
download longtail. if a segment has 65 pieces in the network and the
download request is going to want 39, this selector's job is to turn a
request of 65 nodes into 39 nodes.

this change might not be a good idea but we can try it.

Change-Id: If18de9c7b419ca5fc54869929f32b54ec830cd95