satellite/satellitedb/dbx: add DriverMethods
We have quite a few places where we need access to the dbx methods and
the underlying driver. We have mixed up database and transactions in a
few places. This will allow to avoid some of these problems.

There's also usage in apikeys, to ensure that things work as they
should.

Change-Id: Ifafb461b89ac608138f6882bf90e2f593448e304