storagenode/pieces: flat-file piece expiration store
On many nodes, the piece expiration DB has become very large. Large
tables in databases tend to suffer in performance, particularly when
many queries require a lot of page loads. To make things worse, when
there is a lot of traffic with pieces having expiration times, there is
heavy write load as well.

This solution (optionally) replaces the piece expiration DB with a set
of flat files, one per satellite per hour in the future. (Only hours
that actually have expiration times in them, not every possible hour in
the future, because that would be lol).

Each file is append-only; as pieces come in with expiration times, we
write the piece ID and size to the file. When the hour in question has
elapsed, reading the file to get all newly-expired piece IDs is trivial
and imposes a very light load. Then, removing those expirations from the
piece expiration store is also trivial (we just remove the file).

To avoid having too many file handles open, we use the HandlePool LRU
cache to hold the current set. We expect incoming expiration times to be
fairly well clustered (for example, if a user usually assigns (now + 30
days) as an expiration time) so an LRU pool should be very effective.

Change-Id: I1ba9851e35f352b3c54814985ac034724cc685e7