storagenode/pieces: collector delete should update usage cache
Due to a recent change to skip v0 pieces to speed up the collector,
the delete wasn't going through the BlobsUsageCache because it
doesn't directly have the DeleteWithStorageFormat method.

This change adds the DeleteWithStorageFormat method to the cache
to update the cache and delete the piece.

https: //github.com/storj/storj/issues/7025
Change-Id: I538831d29f4ab0e51a0a7f3e6c526c89054aa158