storagenode/pieces: run cache persistence loop concurrently with the used space calculation
Before the cache persistence loop is started, the used space scan had
to complete. For most cases, if the scan does not complete and there's
an update to the cache, and the node restarts, the cache data is lost,
and with time makes the usage cache data much more inconsistent with the
actual usage.

This change ensures the cache is persisted even while the used space scan
is in progress.

Also fixes a flaky test in TestFilewalker_Basic

https://github.com/storj/storj-private/issues/856
https://github.com/storj/storj/issues/7025

Change-Id: I9b431d5839ee1d01c0f5f5c4ff70be7f2c674e66