testsuite/playwright-ui: Fix flakiness in upload/download
The primary issue is that the success notification obscures the preview
close button when it is displayed, which can cause a timeout during the
test.

This change also updates the id properties for several close buttons in
the UI so that they are unique - and we can be sure that the correct
close buttons are being clicked at different stages in the tests.

Change-Id: I2ab83f69e1864fa3d4678e2553180456ea8efcb1