storagenode/orders: map of open orders files
We currently only keep the most recently used orders file open. If an
order comes in from a different satellite, or was created in a different
window, the currently open orders file is closed and the relevant orders
file for the new order is opened. This can lead to frequent and
unnecessary opening and closing of files.

Instead of only keeping the most recently used orders file open, keep a
map of open orders files. Once the orders from the file are ready to be
sent, close the file and delete it from the map.

issue: https://github.com/storj/storj/issues/7024

Change-Id: I8e841f973e903551318b2126e0b28e5337fdb905