satellite/metainfo: allow object lock if versioning is enabled
This change updates object lock validation from being based to a
project's default versioning. It instead allows object lock to be
enabled, along with versioning on a bucket if;
(1) global versioning is enabled
(2) the project is part of the versioning closed beta
(3) the project is part of the versioning open beta.

Issue: https://github.com/storj/storj/issues/7061

Change-Id: Ib14a3c73298cbbb8a467ca4979d405a3ca9bb5dc