satellite/metabase: drop object lock precommit modes
Main PrecommitDeleteMode purpose was to test some ideas about
improving performance of precommit stage. Unfortunately it was not
well underlined in the code. Only default precommit mode should be
used in production at the moment.

This change removes new precommit mode for object lock and moves logic
to respect retention period for unversioned object (in case of bug).

Logic where we first delete objects and segments and then reverting
transaction in case of active retention is not perfect from
performance perspective but this is only theoretical case for disaster
recovery if we will introduce bug where retention is set to
unversioned object.


Change-Id: I3fe68ffaa9a7a66984d98b0eec2a68985dcb2f45