web/satellite: update delete bucket dialog
This change updates the delete bucket dialog to require users to type
'DELETE' to confirm the deletion of a bucket. This change is intended
to prevent accidental deletions of buckets. Also updated is the text,
and added chips for the bucket's size and object count.

Issue: https://github.com/storj/storj/issues/7083

Change-Id: I8f26d28421506941bcea48f252d73246d1214b05