satellite/{metainfo,console}: combine object versioning and lock betas
This change merges configs for object versioning and lock such that
users allowed to use object versioning can use object lock if it is
globally enabled. In object endpoints, it changes the condition to
use retention aware queries from whether the project is in beta to
whether the relevant bucket is locked. This is to prevent adding new
database calls for the project to check if it is in beta.

NOTE: This change does not consider the eventual consistency of changes
in the object lock status. If a bucket's object lock status is changed,
this change may cause deletion of locked objects on machines that have
not yet seen the change in object lock status. Though we do not yet
support updating a bucket's object lock status, it is important for
this change to be reviewed once again when we do.

Issue: https://github.com/storj/storj/issues/7076

Change-Id: I24abe380ffe706cb772a1f45904b73b0ab06bdf5