web/satellite: update state where limit update and no limits UI are allowed
This change fixes an issue where paid tier users invited to free tier
projects would incorrectly see that the project has no limits. The
change updates the feature to depend on whether the owner of the
project is paid. Also, more restrictions is added to which project
members can update limits; project owners and project admins.

Change-Id: I06cc85c7fd34241a00b1b3e4158b0651526e8fdb