satellite/metainfo: don't check download limits for some projects
This is a bit experimental change to check throughput while download.
Bandwidth limits checks might be unneeded in some cases so would be
good to be able to disable it for some projects.

We don't have explicit definition which limit value will mean
"unlimited" bandwidth so for now I used "-1" and max int64.

Change-Id: I474f4c5c7ad9c247801295ff5446cd3afa2c16a3