satellite/metabase: fix data race when deleting locked objects
This change fixes data races that allowed an object with an active
Object Lock retention period to be deleted if the retention period was
set during the processing of the deletion request.

This issue could have occurred when deleting an exact version of an
object or when deleting the last committed version of an object that
resided in an unversioned bucket. (The latter case isn't expected to
occur because we don't allow Object Lock settings to be specified when
uploading into unversioned buckets, and we don't allow a bucket's
versioning status to be changed if it has Object Lock enabled. However,
we handle this case for safety's sake.)

References storj/edge#461

Change-Id: I4c107467dc548feb46c399616557f32a0e8b8db5