web/satellite,satellite/console: Update managed encryption behavior
Add an additional flag to "project config" to indicate whether managed
encryption is enabled for the project (i.e. a non-empty encrypted
passphrase exists in the DB for the project).

This flag is used to deliberately "break" functionality in the UI if
there is ever a situation where the satellite is unable to decrypt the
project passphrase for a managed-encryption project. Otherwise, the user
of such a project would see a passphrase prompt, the UI would function
without issues, but the user would be unable to download
already-uploaded files, and newly-uploaded files would be uploaded with
the incorrect passphrase.

This change also begins sending Segment analytics events if any
managed-passphrase-related issues occur on the backend. This will allow
us to monitor and debug potential issues.

https://github.com/storj/storj/issues/7097

Change-Id: I646e8cdbe2a81ce6495a64f5fc0496a62a859890