satellite/metainfo: return descriptive error for locked object deletion
This change makes the metainfo service return a descriptive error
message for object deletion requests that fail due to Object Lock
restrictions. Previously, the error message was generic, matching the
behavior of S3 at the time. S3 has since changed, so this change aligns
us with the latest S3 behavior.

Change-Id: If5202a44aea910ff3831df1c7a95c3c6b00f8046