Makefile: generate go1.20 modfile for storagenode
We need to bump go-sql-spanner, which makes the minimum supported Go
version to 1.21; however that's only relevent for satellite.

It would be possible to maintain a fork for go-sql-spanner, however that
requires active maintenance.

The alternative is to create a go.mod file with older version
specifically for Windows storagenode. This is a fragile approach,
because many other things may end up accidentally bumping the version as
well.

Change-Id: I51c80dd5ce7bc2cad1199c934266829cdde01b26