cmd/uplink: avoid materializing recursive copy iterator
this makes it so that if you want to recursive copy millions
of files, it doesn't store an op for every file in a slice,
reducing peak memory usage. it does mean that the output is
a bit less nice (no total files to be copied printed, the file
names are not padded over so harder to parse the output).

we can add those back in for cases where there are a small
number of items from the list iterator by doing a peek if
desired.

Change-Id: Ia3db02d0ca493c2e409fc0a0d6f4bcdea4b54107