storagenode/pieces: read+delete piece expirations from both stores
Prior to this change, if the flat-file piece expiration store was
enabled, any prior existing sqlite3 piece expiration DB was ignored
entirely.

With this change, all piece expiration reads and deletes will be passed
on to the sqlite3 piece expiration DB. Writes will _not_ be passed on,
in particular, which should cause the sqlite3 db to grow smaller and
faster over time.

Change-Id: I758b2dbcdefb712a770b7c976bc9ca8ec5d2ecc1