satellite/metainfo: add test for GetObjectRetention with delete marker
This change adds a unit test verifying the functionality of the
GetObjectRetention metainfo endpoint when used to retrieve the
retention configuration of a delete marker.

Additionally, this change updates the error message that is returned
when the Object Lock settings of a delete marker are requested via the
GetObjectRetention and GetObjectLegalHold metainfo endpoints. This was
done to avoid leaking object information that the requester may not
have permission to see.

Change-Id: I2a278d6e4be98d612be8fba1d149200740abea41