storagenode/trust: Update trusted satellites URL
Update storagenode config to use static.storj.io/dcs-satellites instead
of www.storj.io/dcs-satellites
The www subdomain will continue to serve dcs-satellites for now, but
eventually it will redirect to the static.storj.io subdomain.
A storagenode is still able to fetch the file with the redirect in place,
but it's good to change the config now so that a redirect is not needed
in the future.

Change-Id: I7b59d7383c3def1dc26d2d5761780edef7b38210