satellite/metainfo: reduce BeginSegment response size
BeginSegment response is one of biggest responses in metainfo service.
We would like to reduce it size. This change is easiest step that we can
do right now.

BeginSegment response contains SegmentID which contains duplicated
order limits. Not all fields from those duplicated entries are used
later so we can remove them. One of the fields is SatelliteSignature,
we can remove this one because SegmentID which contains those limits is
already signed by satellite.

With bigger refactoring we may be able to remove more of them but for
now this is minimal set to reduce response by ~15%.

https://github.com/storj/storj/issues/7136

Change-Id: I6df71ea9670972dd15e62839409796c0e3c96c85