satellite/metabase: update error when getting OL info of pending object
This change updates the error information that is returned when a
pending object's Object Lock information is requested via the
GetObjectExactVersionRetention and GetObjectExactVersionLegalHold
metainfo DB methods. Previously, the error indicated that the object
could not be found. Now, the error indicates that the object's state is
incompatible with the operation.

Change-Id: I18d44eeaad2668988c0f635c315b2223467e76b9