satellite/{metabase,metainfo}: update SetObjectLegalHold delete marker error
This change updates the error that is returned when a delete marker's
legal hold status is requested via the SetObjectLastCommittedLegalHold
metabase DB method. Previously, the error indicated that the object
couldn't be found. Now, the error indicates that the object's state is
incompatible with the operation.

Additionally, the error information returned from the
SetObjectLegalHold metainfo endpoint has been updated to be consistent
with the other endpoints that set or retrieve objects' Object Lock
settings.

Change-Id: Ifd19ee99655e5ad1f2157345373fd72664a6ed3e