web/satellite: add tokens step to account setup
This change adds the AddTokens step from the upgrade dialog to the
account setup dialog. A tooltip is added to the "Start Pro Account"
button on the account type step, which allows the user to choose
between adding a credit card or adding STORJ tokens.

Issue: https://github.com/storj/storj/issues/7145

Change-Id: I1d75115cace9816958990babdd4e80edd4d2096e