satellite/nodeselection: new set of selectors for safer choiceofn
choiceofn selection selects n times of the required nodes, and tries to reduce it with pairwise comparison

This method has one problem: it breaks existing invariants. If the original selector guarantees some kind of balanced placement, it can be broken by the choiceofn.

This patch introduces a new choiceofnselection. This one evaluate (=score) each selection (each of them passes the invariant of the selector), and returns with the best one.

Change-Id: I2178d2fc76050f7c3680ca75b8b7c61c867531a0