shared/dbutil: retrying spanner transactions on conflict
We have logic to retry transactions that encounter some conflict, but
the logic did not apply to Spanner. This commit adds the retry logic
needed for Spanner to the transaction helper. It may also be necessary
to add it to outside-of-transaction queries as well, as we do with
`storj.io/storj/shared/dbutil/cockroachutil.Driver`, but we don't have
good evidence either way yet.

Change-Id: I8df48bd5e207981e9bc4d3811654407d68b6dbba