satellite/metabase: clear columns on segment override while upload
There is a case when while committing segment we are overriding previous
versions of this segment. Current problem is that if we are
doing this override we are leaving some columns filled with
remote/inline data.

This change is cleaning columns unused by remote or inline segments
and avoid problem with detecting type of segment.

Change-Id: Ie51c1b20691e60d64cdec7d6b012ee6c22758496