satellite/gc/bloomfilter: support multiple segment sources
We would like to run GC against multiple sources (DBs). To do this
safely we are adding safe check logic for defining bloom filter
creation date:
* get GC start timestamp
* get timestamp of latest created segment from each source
* from all timestamps choose oldest one and use it as bloom filter
creation date

Doing map lookup (`latestCreationTime`) on each segment processing is
not optimal but we decided to do this in this simpler form to avoid
issues with this change initially. Later we will measure impact on GC
process and optimize if needed.

Change-Id: I86e208dce6028f91f0c74418835dd6bee3e8882d