satellite/nodeselection: variant of balanced selector, which supports filtering
filter part of the placement applies to all the actions (upload/download/repair)

If node is filtered out -- eventually -- repair will move it out.

But there are cases when we would like to filter out nodes from upload only (in case of load or anything else).

This is possible with normal random, but will be possible with balanced, only after this patch.

Change-Id: Ic0b52dabb95ecc470b3525fbdb93d608c96e37a7