cmd/storagenode-updater: add --binary-store-dir flag to store current binaries for docker setups
With the current change to ensure binaries persist between restarts,
there are issues executing the binaries from the host mounted location
for setups with lockdown filesystem that do not allow executables.
This flags allows the updater to store a copy of the binary on the
filesystem; on docker restart, this will be copied and executed
instead of executing directly from the host filesystem.

See https://forum.storj.io/t/storagenode-docker-update-changes-to-binaries-location-for-persistence-between-restarts/27937\?u\=clement

Change-Id: I680421fdf315488df33adfe43391c35c57747992