storagenode/hashstore: hash table based store for pieces
It's time for a new approach to the storage node piece storage.

This change introduces a new piecestore backend called hashstore. It has
much in common with AWS' ShardStore. Here's how it works:

Data is stored in log files (called extents in ShardStore). Each log
file is an append only file where piece data is written into it
sequentially. When a new piece comes in, an open and available log file
handle is selected and claimed and the piece is written into that log
file. Then the log file is flushed to disk and returned. If the piece
data is canceled, the log file is rewound to the beginning of that piece
and returned.

The basic index to keep track of which log file and which offset that
pieces are stored in is a disk-backed hash table. Hash table records are
64 bytes long. Each record contains the piece id, the creation
timestamp, the expiration timestamp if any, a flag of if it is trash, a
reference to the log file, an offset within the log file, the length of
the raw piece data, and a checksum.

The basic hash table index and collection of log files is called a
Store. A Store can create new entries, read existing entries, and
compact, which is essentially rewriting the backing hash table index and
log files. This may be needed for garbage collection, trash collection,
hash table resizing, or other compaction and cleanup.

The high level hashstore maintains two Stores, a primary and a
secondary, and swaps between the two when one is undergoing compaction.
Reads are issued to both Stores, and writes are issued to the active
Store.

This design is robust and has few moving parts. Some important points
about this design:
 * The hash table index record is self contained, and is considered
   valid if the checksum in the record matches the record data. If the
   checksum does not match due to corruption, the record is considered
   empty. This means that corruption in one record will not necessarily
   affect any other record.
 * The hash table data structure itself is small and should fit easily
   into operating system caches.
 * Collisions in the hash table are solved with linear probing which
   allows for good performance with sequential disk activity.
 * Garbage collection and trash collection are straightforward
   applications of the standard compaction call and will be very fast
   due to sequential data read and write.
 * Used space calculation will be extremely fast and straightforward,
   due to a very small number of log files.
 * Trash and garbage calculation will be very fast because it can be
   probabilistic based on hash page reads.
 * Trash and expiration times are handled in roughly the same way but
   with an indicator flag of which type of expiration the file is
   subject to (1 week trash policy, or uploader-supplied TTL)
 * Restore from trash is handled very efficiently - the piecestore will
   persist a timestamp-based counter of the last call to restore from
   trash, and will not garbage collect trash that expires within the
   garbage collection window before that timestamp. Compaction will
   clear trash expiration times prior to the end of the garbage
   collection window on next run.
 * Writes to the hash table are random access writes, but in practice
   the records are small and the filesystem is able to help a lot.
   Benchmark performance against BadgerDB shows a ~2x speed improvement
   for many filesizes.

The migration plan will come in a separate PR that integrates the
hashstore with the storagenode.

See benchmarks below

                     │  hashstore  │                badger    │
                     │   sec/op    │    sec/op      vs base   │
Store/Create/0B-32     2.866µ ± 3%    6.182µ ±  5%  +115.68%
Store/Create/512B-32   3.481µ ± 1%    7.162µ ±  4%  +105.76%
Store/Create/1KB-32    3.831µ ± 2%    8.150µ ±  2%  +112.77%
Store/Create/4KB-32    5.529µ ± 1%   17.361µ ±  3%  +214.04%
Store/Create/64KB-32   41.32µ ± 1%   137.33µ ± 11%  +232.35%
Store/Create/1MB-32    617.3µ ± 1%   1430.3µ ±  2%  +131.69%

Change-Id: I7be640f3b7ce876f6961dd20c7405ad1e7bb51e6